+++ lib/src/controls/better_player_cupertino_controls.dart
@@ -496,7 +496,8 @@ class _BetterPlayerCupertinoControlsState extends BetterPlayerControlsState<Bett
           _hideTimer?.cancel();
         },
         onDragEnd: _startHideTimer,
-        onTapDown: cancelAndRestartTimer,
+        onTapDown: _controlsConfiguration.seekOnProgressBarInteractionEnd ? _hideTimer?.cancel : cancelAndRestartTimer,
+        onTapEnd: _controlsConfiguration.seekOnProgressBarInteractionEnd ? _startHideTimer : null,
         colors: BetterPlayerProgressColors(
           playedColor: _controlsConfiguration.progressBarPlayedColor,
           handleColor: _controlsConfiguration.progressBarHandleColor,